+++ CHANGELOG.md
@@ -4,7 +4,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ## Unreleased
 - Parent PLUS loans separated from other family contributions
 - fix migrations for deployment, adding 0003
--
+- Perkins loans hidden for schools that do not offer them
 
 ## 2.1.1
 - Parent PLUS loans separated from other family contributions