+++ CHANGELOG.md
@@ -2,6 +2,16 @@ All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
+- Parent PLUS loans separated from other family contributions
+- fix migrations for deployment, adding 0003
+- Perkins loans hidden for schools that do not offer them
+
+## 2.1.1
+- Parent PLUS loans separated from other family contributions
+- Fixed migrations for deployment, adding 0003
+- Adjusted setup.py to include fixtures and templates
+
+## 2.1.0
 - Pinned and shrinkwrapped NPM dependencies
 - Added `snyk` to monitor for known NPM package vulnerabilities.
 - Updated the project to Django 1.8 and removed solr dependencies
@@ -17,9 +27,12 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Added a 'settlement_school' field for the Student model
 - Updated format-usd to 1.0.1
 - Removed the '.json' suffix from the API call /api/school/[SCHOOL ID]/
+- Added an interaction between Step 1 and Step 2
+- Added school filtering for the Django admin
 - Removed national average graph data & explanatory boxes for settlement schools
 - Added an interaction between Step 1 and Step 2
 - Added school filtering for the Django admin
+- Fix handling of ownership and control values in update_colleges script
 - Added anticipated total direct cost to verification section
 - Fixed file path for BLS json import in views